통계 페이지 수정중
