+++ src/App.jsx
@@ -14,7 +14,8 @@ import FindPW from "./pages/FindPW";  // 비밀번호 찾기
 import SignUp from './pages/SignUp'; // 회원가입
 import SetPW from './pages/SetPW'; // 비번설정 페이지 추가
 import NewPW from './pages/NewPW'; // 새로운 비번 바꾸기 페이지지
-import Statistics from './pages/Statistics'; // 통계
+import StatMain from './pages/Statistics/StatMain'; // 통계
+import Statistics from './pages/Statistics';
 import ProfileSetting from './pages/ProfileSetting'; // 프로필 설정
 import User from './pages/User'; // 마이페이지
 import UserMain from './pages/User/UserMain'; // 마이페이지 메인
@@ -85,7 +86,10 @@ function App() {
           
           
           {/* 통계 페이지 */}
-          <Route path="/statistics" element={<Statistics />} />
+          <Route path="/statistics" element={<Statistics />}>
+            <Route index element={<StatMain />} />
+          </Route>
+
           {/* 마이페이지 */}
           <Route path="/user" element={<User />}>
               <Route index element={<UserMain />} />
