+++ src/components/common/NavBar/index.jsx
@@ -3,10 +3,10 @@ import * as S from "../../../styles/common/NavBar/style.js";
 import DateSelector from "./DateSelector/index.jsx";
 import AlarmIcon from "./AlarmIcon/index.jsx";
 
-const NavBar = ({ icon, modalTop }) => {
+const NavBar = ({ icon, marginTop, modalTop }) => {
 
     return (
-        <S.NavBarWrapper >
+        <S.NavBarWrapper $marginTop={marginTop}>
             <DateSelector modalTop={modalTop}/>
             {icon && <AlarmIcon icon={icon} />}
         </S.NavBarWrapper>
