+++ src/components/common/TopBar/index.jsx
@@ -1,4 +1,4 @@
-import React, {useState} from 'react';
+import React, { useState } from 'react';
 import { TopContainer, LeftTopButton, RightTopButton, TopText, LeftTopUnderline, RightTopUnderline } from '../../../styles/common/TopBar/style';
 import underline from '../../../assets/icons/common/Hide-Indicator.svg';
 
@@ -24,13 +24,13 @@ const TopBar = ({ leftText, rightText, onTabClick, selectedTab }) => {
       <TopContainer>
         {/* 왼쪽 버튼 */}
         <LeftTopButton onClick={() => handleTabClick(0)}>
-          <TopText isSelected={activeTab === 0}>{leftText}</TopText>
+          <TopText $isSelected={activeTab === 0}>{leftText}</TopText>
           {activeTab === 0 && <LeftTopUnderline src={underline} alt="underline" />}
         </LeftTopButton>
   
         {/* 오른쪽 버튼 */}
         <RightTopButton onClick={() => handleTabClick(1)}>
-          <TopText isSelected={activeTab === 1}>{rightText}</TopText>
+          <TopText $isSelected={activeTab === 1}>{rightText}</TopText>
           {activeTab === 1 && <RightTopUnderline src={underline} alt="underline" />}
         </RightTopButton>
       </TopContainer>
